refactor(prebuilt): :recycle: extracted LeaveRoom and EndRoom modals out of `HMSManageLeave` component
